BugFix: Support for correct runner.total count when using --grep
When running --grep the runner.total is updated to the total number of
tests matched by the grep search.

Added an .iterateTests() method for the Suite class that iterates
recursively over all tests within the suite and applies a function.
This is used to get the total number of tests that match grep search.

In order to apply .iterateTests() before the reported is called it was
needed to move the 'new Reporter()' call to after the 'runner.grep()'
call in 'bin/_mocha'.