Fix leak of listeners for process
When I run `mocha --watch` and after 11th running, the following message was
shown:

```
(node) warning: possible EventEmitter memory leak detected. 11 listeners
added. Use emitter.setMaxListeners() to increase limit.
```

It seems that 'uncaughtException' listener is not released.
Because, the listener to be released is different from listened one.

To fix this issue, I gave the same listener to `process.on` and
`process.removeListener`.