Add integration mode to html reporter
When adding tests to an existing page, we want
mocha to be unobtrusive. Adding the css class
`integration` to the mocha div will enable the
integration mode of the html reporter. This will:

* Remove the report from the page flow, and hide
it by default.
* Add a toggle to the stats panel to show/hide the
report.
* Modifies the css to remove some built in 
assumptions, where the report styles would be
inherited from the base page. (and thus unreadable
if it is loaded in a page with a white on dark
color scheme, for example).
* Add the css class `bottom` to the mocha div, and
the stats panel will move to the bottom of the page.
* Adds a test to demonstrate its usage. Run:
`node test/browser-integration/index.js`, and visit
`http://localhost:3333/`