Separate initialization of mocha with `run`
Motivation: I have written an adapter for testacular
that will load a page with mocha tests inside an
iframe. I need to be able to hook (but not change)
the runner events to report the status of the tests
to testacular. Thus:

```
var mocha = iframe.contentWindow.mocha;
var runner = mocha.init();
runner.on('start', function() {
  tc.info({total: runner.total});
});
runner.on('test end', reportTest);
runner.run();
```