Minor fix to HTML reporter display.
Previously the HTML reporter displayed 'pre' blocks with display set
to 'inline-block', however given a wide enough screen, when a test
fails and you click the h2, the code snippet will display beside the
error output, as opposed to below it which would look more
consistent. This commit fixes this styling.