Preserve all query params expect `test` and `grep` for grep links
For HTML reporter.

There are webapps that change behavior depending on the query
parameter values. For example, this demo application chooses its
engine based on the value of the `engine` query parameter:

http://tkareine.org/hackjkl-intro-frp/formapp/index.html?test&engine=bacon

Without this change, when we click on a specific test context, we lose
engine selection:

http://tkareine.org/hackjkl-intro-frp/formapp/index.html?grep=After%20loading%20application%20when%20entering%20space%20for%20search

This change preserves all query parameters except `test` and `grep`
when creating grep links in HTML reporter.