Add globbing to support windows-style globs.
In unix shell, globs are expanded by the shell, whereas in windows
globs are passed to the application for expansion.

If the specified filepath doesn't exist and filepath.js doesn't exist,
then try globbing that string.  This has no runtime cost in most
scenarios, and if filepath is not a glob the result is unchanged.

Closes #844