Add a feature to optionally direct the test output to a file.
The implementation creates a new output library and changes all of the
test reporters to call output.log or output.write instead of directly
calling console.log or process.stdout.write.

Added a test rule to the Makefile that lists all of the reporters and
verifies that the output is properly generated when the option is
enabled.