Tweak file output implementation based on pull request comments.
Pass the output stream to each of the reporters in the constructor to make
integration simpler for reporters defined outside of the mocha source tree and
to simplify the changes.

Also add support to pass an arbitrary stream to output.initialize and rename
the 'stream' variable to 'ostream' so it doesn't clash with the stream module.