Correct erroneous test
Because the Runnable under test is not part of a Suite, its context is
undefined. Prior to this commit, the error returned was a
runtime-generated `ReferenceError`, not a Mocha-generated `Error`.

Set the timeout via the `Runnable#timeout` method, and update the test
to assert the correct behavior (not implementation details such as the
message body of the generated error).