Fixed Broken Nyan Reporter
The Nyan reporter is the only reporter that modified the reporter
prototype. This prototype is overwritten in the mocha.js compilation,
causing the reporter to fail when used in mocha.js. Normally mocha.js
is only used in browsers, so this reporter wouldn't work anyway. But
there are many CLIs that use mocha & phantomJS to run front-end tests
on the command line. Because they rely on mocha.js, nyan wasn't working.

This commit removes the prototype modifications and moves functions off
of the NyanCat instance, insted binding them to be run with the correct
this context. This way the prototype is  unmodified, and the Nyan reporter
now works in mocha.js.