#888 Fix for --watch+only() issue
When .only() is used mocha.grep is set to the title of the .only test. This means that if --watch is on and it reruns the test and the .only has been removed the mocha.grep is still set to the same string. We clear mocha.grep on rerun and that means if .only has been moved or remains mocha.grep will be set again or remain cleared.