reporters/base: use supports-color to detect colorable term
This replaces https://github.com/mochajs/mocha/pull/1497.

Detecting whether a terminal is TTY is not enough
to ensure that it can also support color.
This module, supports-color, is a wholistic solution
which captures the greatest range of possible terminals
with the greatest accuracy. The module is extremely
lightweight and does not represent a burdern to Mocha
users.