Stricter .only for suites: suites whose names are a superset of the .only suite don't wind up also running.
Similar to 179ffd35, an earlier fix for .only on tests.

Limitation: because test names are concatenated to suite names, the grep
can't be AS strict as it is for .only on tests. When the name of the .only
suite + " " is a prefix of another suite's name, that other suite will
still run even with this fix.