build mochajs.org with Jekyll; closes #1640
Lots of stuff here, but the main things:

- use Jekyll
- convert `index.md` to *actual GFM Markdown*
- a lot of edits of the documentation, including, but not limited to:
  - reorganization
  - grammar fixes
  - code style fixes
  - rewording confusing sentences
  - add a couple more examples
  - dynamically generate TOC
  - use CDNs for JS, though there's basically none

I've taken every effort to make the site look as close to the current version as it does now, and keep it compatible with old browsers.  However, I was not able to reuse the syntax highlighting scheme, as highlighting is now done by Pygments.

![screenshot 1](http://content.screencast.com/users/chiller/folders/Jing/media/32a9e9d6-b927-465a-a027-39d8bbfe0407/00000153.png)

![screenshot 2](http://content.screencast.com/users/chiller/folders/Jing/media/890e0481-65a4-4729-bc69-37098c48940f/00000154.png)