Call the inspect() function if message is not set
Unfortunately some frameworks throw error objects which do not have a `message`
property but do have an `inspect()` function. This means Mocha reports a test
failure, but prints the empty string instead of any useful information about
the error message. For an example, see versions of the Waterline ORM before
v0.10.19: https://github.com/balderdashy/waterline/commit/0965d132

If the `message` key is not set on an error object, attempt to print the same
output as `console.log` in Node, by calling the object's `inspect()` function,
if it exists.

We could try to fallback to calling `util.inspect` on the `err` object and
logging that, but I'm not sure the format would be appropriate.