chore(Saucelabs): test on IE9, IE10 and IE11.
Fixes #2344.

This moves the selection of the browsers to test to the travis build
script. The number of concurrent travis jobs should be limited to 5
(Saucelabs limit).
This also removes the S3 env variable which does nothing

This uses a fork of karma-sauce-launcher, which contains upated
dependencies, in order to support retries:
https://github.com/karma-runner/karma-sauce-launcher/pull/114