Fix wrapping behaviors [ci skip]
GitHub wraps in the new issue editting box, so we don't have to; we should put some space around that URL since it's expected to wrap, though. Alternatively if we are to manually wrap at all we'd need to wrap everywhere consistently. Before this PR the areas changed by this commit look... odd due to the combination of the hardcoded wrapping and the GitHub wrapping.