execute test after "before each" hook asynchronously; closes #3009
This provides consistent, though not necessarily ideal, behavior when a
`Promise` is rejected within a "before" or "before each" hook and *not*
caught synchronously or within a `process.nextTick()` callback.

Previously, the `Promise` would be considered handled if it was rejected
in a "before each" hook and caught within the test, because the test
would be executed synchronously.  This is not the same case for a
"before" hook, which would consider the `Promise` unhandled.  Now, both
result in an unhandled exception warning from node.

An alternative implementation would be to execute tests within a
`process.nextTick()` callback (not sure what happens in a browser).
This would mean rejected `Promise`s in both cases would be considered
handled, because `process.nextTick()` would execute *before* the
microtask which then considers the `Promise` to be unhandled.

Mocha executes tests recursively, and needs to do so
asynchronously in order to avoid exceeding the maximum call stack.  As
per the Node.js documentation [here](https://nodejs.org/en/docs/guides/event-loop-timers-and-nexttick/),
this is *not* a valid use case of `process.nextTick()`.

Yet another alternative implementation which may avoid a decision
entirely would be to execute tests using a generator.  This could only
be *simulated* in an environment such as IE11, and thus would
potentially result in inconsistent or undefined behavior when running
the same tests in different browsers (that being said, generators are
probably a Good Idea for Mocha to consume in the long run).