Don't issue deprecation error for -compilers with --ignore-compilers
This is useful when a project has multiple test directories, some using older mocha with -compilers and others that don't need a compiler. WebStorm only supports a single global configuration in a project for mocha options, so with this --ignore-compilers option, the older mocha continues to use compilers and the new mocha doesn't throw an error.