Introduce the afterThis hook
The afterThis hook allows one to write cleanup logic inside the test definition,
enabling the use of variables declared inside the test itself. For example:

```js
// without afterThis hook
let resource;
beforeEach(() => {
    resource = acquireResource();
});
afterEach(() => {
    resource.dispose();
});
it('does something with a resource', () => {
    // use `resource`
});

// using afterThis hook
it('does something with a resource', () => {
    const resource = acquireResource();
    afterThis(() => resource.dispose());

    // use `resource`
});
```

The code using the resource is much closer to the initialisation/cleanup logic,
and there are less lifecycle hooks to be aware of.

This is inspired by bash's `trap` and go's `defer`.