Add support for `reporterOptions` as an object
When using a configuration file, users can pass options to a reporter via the `reporterOption` property. The value of this property is expected to be a string or comma-separated list of "key=value" pairs. This format matches the CLI but is a bit awkward in a config file where using an object would be more natural. This commit enables the object format by coercing it to a "key=value" string that can be parsed by yargs.