Remove reference to misleading article and any platforms
https://github.com/mochajs/mocha/pull/4869#pullrequestreview-1931377316 requested that platforms and shells not be mentioned. This means that rationale for the directions cannot be provided.

The article which was originally linked suggests to use single quotes which do not work on Windows. For example, using single quotes for a glob expression referring to a folder or file with `&` in its name would break the expression into two commands on Windows.

Since the rationale was removed and the linked article was misleading, instructions on how to correctly quote globs were added. The rationale is recorded here since it isn’t allowed in the instructions:

* Since a Unix-like OS user would be accustomed to referencing a path containing `"` in it by writing `\"` within a double-quoted string, I mentioned that the double quote and backslash characters are forbidden.
* Since a Windows user would be accustomed to including `$` in file names but double quoted strings have parameter expansion performed on them on on Unix-like OSes, I had to mention that `$` is a forbidden character.
* Since Windows doesn’t support single quotes, I had to mention that double quotes are required.