refactor: replace `fs-extra` with newer `fs` built-ins (#5284)
* Replace fs-extra with newer fs built-ins

This updates the codebase to remove two uses of fs-extra by newer fs
built-in functions. The code changes are based on the implementation of
fs-extra@10.1.0, which is the version of fs-extra recorded in the
project's lockfile.

In particular, for `watch.spec.js`, the changes are based on:
https://github.com/jprichardson/node-fs-extra/blob/0220eac966d7d6b9a595d69b1242ab8a397fba7f/lib/remove/index.js#L15

and, also for `helper.js`, the changes are based on:
https://github.com/jprichardson/node-fs-extra/blob/0220eac966d7d6b9a595d69b1242ab8a397fba7f/lib/mkdirs/make-dir.js#L23

* Drop 14 and 16 in mocha.yml

---------

Co-authored-by: Josh Goldberg <git@joshuakgoldberg.com>