fix: print error when internal cli error happens
There have been issues where the reported error is "ERROR: null", which
is very unhelpful. Here is an analysis of a specific example:
https://github.com/mochajs/mocha/issues/5048#issuecomment-2833619212

Although the trigger for that error was fixed by #5074, the
unhelpfulness of "ERROR: null" was not addressed.

To help with debugging, this patch prints the original error when this
stage is unexpectedly reached.