`lookupFiles` repeats various operations over and over. We can avoid alot of this overhead and improve the performance of big test suites.
To improve the performance, we normalize the file extensions once at the beginning and dont need to do it again.

This means that we can avoid again normalizing the array in `hasMatchingFileExtension` (before: `hasMatchingExtname`). Also we avoid the performance bottleneck of Array.helper functions like `.map` and `.some()`.

`fs.statSync` is called with `throwIfNoEntry`, which exists since node 14.17.0, set to `false`. We can avoid "expensive" errors and the try catch blocks. We just check if stat is undefined, and if so, well, the path has an issue.

Using `recursive` and `withFileTypes` option of readdirsync, we can avoid using recursive calling of `lookupFiles`.

Added `jsdoc` for better type resolution. Also if you would put `// @ts-check` at the beginning of the file, it would pass.